完成菜单增删改查拖动逻辑

+++ database/migrations/2016_04_26_111015_create_menus_table.php
@@ -20,7 +20,7 @@ class CreateMenusTable extends Migration
             $table->string('name', 50)->comment('菜单名称');
             $table->string('route', 50)->comment('菜单路由');
             $table->string('description', 50)->comment('菜单描述');
-            $table->tinyInteger('sort')->default(0)->comment('菜单排序');
+            $table->integer('sort')->default(0)->comment('菜单排序');
             $table->tinyInteger('hide')->default(0)->comment('是否隐藏');
             $table->timestamps();
         });
